Move page not found to global folder
