+++ src/app/app-routing.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { NewRequestsComponent } from './new-requests/new-requests.component';
 import { UserSettingsComponent } from './user-settings/user-settings.component';
-import {PageNotFoundComponent} from "./page-not-found/page-not-found.component";
+import {PageNotFoundComponent} from "./global/page-not-found/page-not-found.component";
 
 const routes: Routes = [
   {path: 'premium/requests', component: NewRequestsComponent},
