+++ src/app/app.module.ts
@@ -15,7 +15,7 @@ import { LayoutModule } from '@angular/cdk/layout';
 import { MatSidenavModule } from '@angular/material/sidenav';
 import { MatIconModule } from '@angular/material/icon';
 import { MatListModule } from '@angular/material/list';
-import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
+import { PageNotFoundComponent } from './global/page-not-found/page-not-found.component';
 
 
 function initializeKeycloak(keycloak: KeycloakService) {
